+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+- 2022-12-17: added `Id` to `SearchResult` for better handling of search results by client code.
+
 ## [0.0.4] - 2022-12-15
 
 - 2022-12-15: updated packages.